[#159742926] Respond to PR feedback
Move multi date input to own file
Specify Number on date input fields
