+++ app/javascript/utils/form/Field.js
@@ -5,7 +5,7 @@ import { Select, Text } from 'react-form'
 
 import { FormError, errorClassName } from './errors'
 
-const BlockNote = ({ value }) => (
+export const BlockNote = ({ value }) => (
   <span className='checkbox-block_note no-margin'>{value}</span>
 )
 
@@ -20,31 +20,6 @@ export const Field = ({ formApi, field, label, blockNote, errorMessage, children
       </label>
       { children(field, classNames(className)) }
       <FormError formApi={formApi} label={label} field={field} errorMessage={errorMessage} />
-    </div>
-  )
-}
-
-export const MultiDateField = ({ formApi, field, label, id, blockNote, errorMessage, children }) => {
-  const className = errorClassName(formApi, field)
-
-  return (
-    <div className={classNames('form-group', className)}>
-      <label className='form-label' htmlFor={field}>
-        {`${label} `}
-        {blockNote && <BlockNote value={blockNote} />}
-      </label>
-      <div className='form-group-month'>
-        <Text className={classNames(className)} field={[field, 1]} id={id + '_month'} placeholder='MM' />
-      </div>
-      <div className='form-group-day'>
-        <Text className={classNames(className)} field={[field, 2]} id={id + '_day'} placeholder='DD' />
-      </div>
-      <div className='form-group-year'>
-        <Text className={classNames(className)} field={[field, 0]} id={id + '_year'} placeholder='YYYY' />
-      </div>
-      <div className='d-inline-block'>
-        <FormError formApi={formApi} label={label} field={field} errorMessage={errorMessage} />
-      </div>
     </div>
   )
 }
